+++ src/cli/config.rs
@@ -1,10 +1,14 @@
 use anyhow::{bail, Result};
-use clap::Parser;
+use clap::{Parser, ValueEnum};
 use colored::Colorize;
 use open;
-use std::fs::{self, File};
+use std::{
+	env,
+	fs::{self, File},
+	path::PathBuf,
+};
 
-use crate::{argon_info, config::Config as GlobalConfig, ext::PathExt, logger, util};
+use crate::{argon_info, config::Config as ArgonConfig, ext::PathExt, logger, util};
 
 /// Edit global config with default editor or CLI
 #[derive(Parser)]
@@ -24,6 +28,14 @@ pub struct Config {
 	/// Restore all settings to default values
 	#[arg(short, long)]
 	default: bool,
+
+	/// Which config file to work with (`global` or `workspace`)
+	#[arg(short, long, hide_possible_values = true)]
+	config: Option<ConfigType>,
+
+	/// Save current config to the custom file
+	#[arg(short, long)]
+	output: Option<PathBuf>,
 }
 
 impl Config {
@@ -31,18 +43,29 @@ impl Config {
 		if self.list {
 			argon_info!(
 				"List of all available config options:\n\n{}\nVisit {} to learn more details!",
-				GlobalConfig::list(),
+				ArgonConfig::list(),
 				"https://argon.wiki/docs/configuration#global-config".bold()
 			);
 
 			return Ok(());
 		}
 
-		if self.default {
-			let config_path = util::get_argon_dir()?.join("config.toml");
+		match self.config.unwrap_or_default() {
+			ConfigType::Global => ArgonConfig::load_global(),
+			ConfigType::Workspace => ArgonConfig::load_workspace(&env::current_dir()?),
+			_ => {}
+		}
 
+		let config = ArgonConfig::new();
+		let config_path = if let Some(path) = config.kind().path() {
+			path.to_owned()
+		} else {
+			util::get_argon_dir()?.join("config.toml")
+		};
+
+		if self.default {
 			if config_path.exists() {
-				if GlobalConfig::new().move_to_bin {
+				if config.move_to_bin {
 					trash::delete(config_path)?;
 				} else {
 					fs::remove_file(config_path)?;
@@ -54,16 +77,19 @@ impl Config {
 			return Ok(());
 		}
 
+		let output = self.output.unwrap_or(config_path);
+
 		match (self.setting, self.value) {
 			(Some(setting), Some(value)) => {
-				let mut config = GlobalConfig::new_mut();
+				drop(config);
+				let mut config = ArgonConfig::new_mut();
 
 				if config.has_setting(&setting) {
 					if let Err(err) = config.set(&setting, &value) {
 						bail!("Failed to parse value: {}", err);
 					}
 
-					config.save()?;
+					config.save(&output)?;
 
 					argon_info!("Set {} to {}", setting.bold(), value.bold());
 				} else {
@@ -71,16 +97,17 @@ impl Config {
 				}
 			}
 			(Some(setting), None) => {
-				let default = GlobalConfig::default();
+				let default = ArgonConfig::default();
 
 				if default.has_setting(&setting) {
-					let mut config = GlobalConfig::new_mut();
+					drop(config);
+					let mut config = ArgonConfig::new_mut();
 
 					config
 						.set(&setting, &default.get(&setting).unwrap().to_string())
 						.unwrap();
 
-					config.save()?;
+					config.save(&output)?;
 
 					argon_info!("Set {} to its default value", setting.bold());
 				} else {
@@ -88,24 +115,37 @@ impl Config {
 				}
 			}
 			_ => {
-				let config_path = util::get_argon_dir()?.join("config.toml");
-
-				if !config_path.exists() {
-					let create_config = logger::prompt("Config does not exist. Would you like to create one?", true);
+				// TODO: fix this when there is workspace config
+				if !output.exists() {
+					let create_config = logger::prompt(
+						&format!(
+							"{} config does not exist. Would you like to create one?",
+							config.kind().to_string().bold()
+						),
+						true,
+					);
 
 					if create_config {
-						File::create(&config_path)?;
+						File::create(&output)?;
 					} else {
 						return Ok(());
 					}
 				}
 
-				argon_info!("Opened config file. Manually go to: {}", config_path.to_string().bold());
+				argon_info!("Opened config file. Manually go to: {}", output.to_string().bold());
 
-				open::that(config_path)?;
+				open::that(output)?;
 			}
 		}
 
 		Ok(())
 	}
+}
+
+#[derive(Clone, Default, ValueEnum)]
+enum ConfigType {
+	#[default]
+	Default,
+	Global,
+	Workspace,
 }